requirements.txt
dspy-ai
pydantic-ai[logfire]
httpx
openai
beautifulsoup4